Move tableNames.js to model instead of helpers
