+++ src/api.v2/model/Experiment.js
@@ -6,7 +6,7 @@ const { collapseKeyIntoArray } = require('../../sql/helpers');
 
 const { NotFoundError } = require('../../utils/responses');
 
-const tableNames = require('../helpers/tableNames');
+const tableNames = require('./tableNames');
 
 
 const getLogger = require('../../utils/getLogger');
